% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_oce.R
\name{geom_oce}
\alias{geom_oce}
\title{geom_oce}
\usage{
geom_oce(data = NULL, mapping = NULL, lonlim, latlim)
}
\arguments{
\item{data}{The data for drawing a plot}

\item{mapping}{Mapping for aes()}

\item{lonlim}{The lontitude limits: lonlim = c()}

\item{latlim}{The latitude limits: latlim = c()}
}
\value{
A plot
}
\description{
geom_oce() is used for ocean data visualization
}
\examples{
\dontrun{
library(nppr)
library(ggplot2)
library(ggspatial)
library(rnaturalearth)
library(rnaturalearthdata)
data(nppdata)
ggplot()+
  geom_oce(nppdata, aes(x = lon, y = lat, fill = npp),
  lonlim = c(100, 120), latlim = c(7, 25))+
  scale_fill_gradientn(colors = rev(rainbow(20)), breaks = seq(200, 1000, 100),
  limits = c(200, 1000))+
  labs(x = 'Longitude', y = 'Latitude',
  fill = expression(NPP*~'('*mg~C~m^-2~d^-1*')'))
}
}
