% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-npp-samples.R
\name{get.npp.samples}
\alias{get.npp.samples}
\title{Take a (non-probability, probability) pair of samples from the given synthetic population}
\usage{
get.npp.samples(
  DF.population = NULL,
  prob.selection = 0.2,
  n.replicates = 500,
  seed = 7654321
)
}
\arguments{
\item{DF.population}{data frame containing synthetic population data (generated with \code{nppR::get.synthetic.population( )})}

\item{prob.selection}{numeric vector of length 1 containing the selection probability of the SRS probability sample. Must be a number strictly between 0 and 1. Default is 0.2.}

\item{n.replicates}{numeric vector of length 1 containing the number of sets of bootstrap weights. Must be a positive integer. Default is 500.}

\item{seed}{numeric vector of length 1 containing randomization seed. Must be a positive integer. Default is 7654321.}
}
\value{
list containing the non-probability and probability samples as data frames
}
\description{
The \code{get.npp.samples} function draws a (non-probability, probability) pair of samples from the given synthetic population data frame (generated with \code{nppR::get.sythetic.population( )})
}
\examples{
DF.population <- nppR::get.synthetic.population(
    population.size = 10000
    );

list.samples  <- nppR::get.npp.samples(
    DF.population = DF.population,
    prob.selection = 0.2,
    n.replicates   = 500
    );

}
