% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-synthetic-population.R
\name{get.synthetic.population}
\alias{get.synthetic.population}
\title{Generate a synthetic population}
\usage{
get.synthetic.population(population.size = 10000, seed = 1234567)
}
\arguments{
\item{population.size}{numeric vector of lengh 1 indicating desired population size. Must be a positive integer. Default value is 10000.}

\item{seed}{numeric vector of length 1 containing randomization seed. Must be a positive integer. Default is 1234567.}
}
\value{
a data frame containing data of the synthetic population.
}
\description{
The \code{get.synthetic.population} function generate a synthetic population; it is used in the vignettes.
}
\examples{
DF.population <- nppR::get.synthetic.population(
    population.size = 10000
    );

}
