% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_npp_eppley.R
\name{get_npp_eppley}
\alias{get_npp_eppley}
\title{get_npp_eppley}
\usage{
get_npp_eppley(
  file.path,
  grid.size = "low",
  time.span = "monthly",
  file.format = "hdf",
  satellite = "MODIS",
  mindate,
  maxdate
)
}
\arguments{
\item{file.path}{The folder(an empty folder) path where your want to
save your file (avoid Chinese characters).}

\item{grid.size}{The grid size that you choose. There are two grid sizes can be choosed:
'low'(default): 2160x1080, 'high': 2160x4320.}

\item{time.span}{The time span of npp data. There are two time spans:
'monthly' represent monthly npp data. 'dayly' represent 8 days data.}

\item{file.format}{For VGPM model, there are two file formats can be choosed, 'xyz' and 'hdf'. For Eppley-VGPM, CBPM and CAFE, only 'hdf' format can be choosed.}

\item{satellite}{Choose satellites, 'MODIS', 'VIIRS', and 'SeaWiFS'. The default is 'MODIS'.}

\item{mindate}{The minimum date of data you want to download.}

\item{maxdate}{The maximum date of data you want to download.}
}
\value{
download some files in your folder.
}
\description{
get_npp_eppley() is used for automatically downloadiing, decompressing and renaming
ocean net primary production data of Eppley-VGPM model by custom grid size, time
span and satellite.
}
\note{
units: mg C m^-2 d^-1
}
\examples{
\dontrun{
library(nppr)
library(RCurl)
library(XML)
library(R.utils)
library(tidyverse)
library(lubridate)
get_npp_vgpm(file.path = 'C:\\\\Users\\\\xucha\\\\Desktop\\\\DATA',
        mindate = '2016-02-04', maxdate ='2016-06-28')
}
}
\author{
Chao Xu
}
