% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_par.R
\name{get_par}
\alias{get_par}
\title{get_par}
\usage{
get_par(
  file.path,
  grid.size = "low",
  time.span = "monthly",
  satellite = "MODIS",
  mindate,
  maxdate
)
}
\arguments{
\item{file.path}{The folder(an empty folder) path where your want
to save your file (avoid Chinese characters).}

\item{grid.size}{The grid size that you choose. There are two grid sizes can be choosed:
'low'(default): 2160x1080, 'high': 2160x4320.}

\item{time.span}{The time span of npp data. There two time spans:
'monthly' represent monthly npp data. 'dayly' represent 8 days data.}

\item{satellite}{Choose satellites, 'MODIS' and 'VIIRS'. The default is 'MODIS'.}

\item{mindate}{The minimum date of data you want to download.}

\item{maxdate}{The maximum date of data you want to download.}
}
\value{
download some files in your folder.
}
\description{
get_par() is used for automatically downloadiing, decompressing and renaming
sea photosynthetically available radiation data by custom grid size, time
span and satellite.
}
\note{
units: Einstein m^-2 day^-1
}
\examples{
\dontrun{
library(nppr)
library(RCurl)
library(XML)
library(R.utils)
library(tidyverse)
library(lubridate)
get_chla(file.path = 'C:\\\\Users\\\\xucha\\\\Desktop\\\\DATA',
        mindate = '2016-02-04', maxdate ='2016-06-28')
}
}
\author{
Chao Xu
}
