% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_mon_day.R
\name{match_mon_day}
\alias{match_mon_day}
\title{match_mon_day}
\usage{
match_mon_day(data, file.path)
}
\arguments{
\item{data}{Your npp data, it can be a data frame or a tribble.}

\item{file.path}{The folder path where the npp data seved.}
}
\value{
A data frame
}
\description{
match_mon_day() is used to export multiple ocean productivity values from the download files
by your longitude, latitude and date
}
\note{
This function should only be used when you have the 'month' variable.
Any other time, match_sig should be used. Please see examples.
}
\examples{
\dontrun{
library(nppr)
librray(tidyverse)
library(lubridate)
# Your data must contain both 'date' and 'month' variables
mydata \%>\% mutate(month = as.yearmon(format(date, '\%Y-\%m'))) \%>\% # create 'month' variable
group_by(month) \%>\% nest() \%>\%
mutate(var = map(data, ~match_mon_day(data, file.path = 'C:\\\\Users\\\\xucha\\\\Desktop\\\\DATA'))) \%>\%
unnest(var) \%>\% select(date, var))
}
}
