% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hdf.R
\name{read_hdf}
\alias{read_hdf}
\title{read_hdf}
\usage{
read_hdf(file.path, var = "npp")
}
\arguments{
\item{file.path}{The file path of which one you want to import.}

\item{var}{The variable in the file. The default is 'npp'.}
}
\value{
A data frame
}
\description{
read_hdf() can be used to read the ocean productivity(npp, sst, chl and par) hdf file in your R work environment
}
\examples{
\dontrun{
library(nppr)
library(raster)
librray(tidyverse)
read_hdf(data, file.path = 'C:\\\\Users\\\\xucha\\\\Desktop\\\\DATA\\\\cbpm.201506.hdf')
}
}
