% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppr-package.R
\docType{package}
\name{nppr-package}
\alias{nppr}
\alias{nppr-package}
\title{nppr: Tools for Ocean Productivity Data Downloading, Reading, Processing and Visualization}
\description{
Tools for ocean productivity data downloading, reading, processing and visualization. All data are available from <http://sites.science.oregonstate.edu/ocean.productivity/index.php>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chaoxv/nppr}
}

}
\author{
\strong{Maintainer}: Chao Xu \email{xuchao3091@live.com}

}
\keyword{internal}
