% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppr-package.R
\docType{package}
\name{nppr-package}
\alias{nppr}
\alias{nppr-package}
\title{nppr: Ocean productivity data downloading, reading, processing and visualization}
\description{
This package is used for ocean productivity data downloading, reading, processing and ocean data visualization. All data are available from <http://sites.science.oregonstate.edu/ocean.productivity/index.php> or <http://orca.science.oregonstate.edu/data>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/chaoxv/nppr}
}

}
\keyword{internal}
